issue #26 - include parent classes in element definition ids to avoid conflicts
In DSTU2 we prefixed classes with the name of the containing type.
However, in R4, we moved to a nested class structure.
Furthermore, there are now some BackboneElements that share a local name
within the same resource (e.g. ImplementationGuide.Definition.resource
and ImplementationGuide.Manifest.resource) and so we need to use the
full path of contained resource simple names to ensure uniqueness.

Signed-off-by: lmsurpre <lmsurpre@us.ibm.com>