issue #26 - added delimiter between class names in obj defs
Per review comments, I added a "_" separator between the class names
when constructing the field name for object definitions.
This way is slightly safer, just in case there are any cases where names
would not be unique (e.g. if we had cases like Resource.backbone.one.two
and ResourceId.backbone.onetwo).
The "_" is also used in the JSON schema shipped with the specification.

Signed-off-by: lmsurpre <lmsurpre@us.ibm.com>