issue #182 - update FHIRValidator.getResource to handle contained resources properly
Previously, we always returned the root of the FHIRPath tree of nodes for datatype constraints, but this was wrong for contained resources within a Bundle entry.

Also added unit tests for validating various constraints within a Bundle entry, including:
* Bundle in Bundle
* local reference to contained resource within a bundle
* local reference to a peer contained resource within a bundle

Also converted FHIRValidatorTest to be a unit test and removed the unnecessary outputStream wrapper.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>