issue #220 - re-enable JDBCSearchDateTest and fix related issues
Enabling this test as part of issue #92 had uncovered issue #220.
This commit addresses issue #220 by passing FHIR Elements to JDBC
"Processor" component which extracts the values from a node.

Previously, we were feeding the process method the entire FHIRPathNode
and it was processing the start and end times separately (instead of
handling them together as a date range).

To make the JDBCNormSearchDateTest work, I
1. moved the BasicDate.json example to the fhir-examples project
2. added a helper method in the AbstractPLSearchTest to facilitate
reading examples from fhir-examples
3. updated the extension-search-parameters.json config for the "date"
tenant and introduced the newly required
extension-search-parameters-valuetypes.json file
4. moved the `createResources` step into the setup in order to avoid
"dependsOn" and the corresponding issue I opened at
https://github.com/cbeust/testng-eclipse/issues/435 ...this change
allows us to run individual test cases from the test class whereas that
used to fail. To ensure this runs after the @BeforeClass methods being
used to set the tenantId in the concrete subclasses, I added all the
tests to a group and used the @BeforeGroups annotation; perhaps there is
a cleaner way but this seems to work ok.

If this approach is supported by the team, I will proceed to enable the
other search tests in a similar fashion.

Signed-off-by: lmsurpre <lmsurpre@us.ibm.com>