issue #226 - additional cleanup per review
Renamed JDBCParameterBuilder2 to JDBCParameterBuildingVisitor per review comments.

Also modified Visitor to include a visit for any Visitable and changed DefaultVisitor to delegate all visits to the parent.
This replaces the "defaultAction" methods and unifies our approach toward class hierarchy in this class.

And one unrelated change: I fixed a typo in the description of a constraint on InsurancePlan

Signed-off-by: lmsurpre <lmsurpre@us.ibm.com>