issue #92 - move search test config to fhir-persistence
these config files are tightly coupled with the [abstract] search tests
and so they should be in the same project

previously, this config data was in fhir-persistence-jdbc project simply
out of expedience. the downside of this approach is that you need to
bounce between the fhir-persistence-jdbc and fhir-persistence projects
to update the tests. furthermore, we'd need to copy the same config
files to each and every persistence implementation.

the one downside to moving these config files is that I had to introduce
a dependency on the project layout...specifically peristence layer
implementations providing concrete subclasses of these tests must be in
a directory that is peer to the fhir-persistence project (due to
reliance on filesystem paths in FHIRConfiguration).

Signed-off-by: lmsurpre <lmsurpre@us.ibm.com>