issue #92 - refactor SearchUtil.parseQueryParams and enable tests
1. remove queryString from parseQueryParameters methods

Previously, all variants of SearchUtil.parseQueryParameters wanted the
queryString in
addition to the Map of queryParameters. This doesn't make much sense
because the queryParameters map should have everything we need.

Now we can just pass the queryParameters. Changes needed to enable this
include:
* removal of queryStringIndex from SortParameter (just pass them in the
right order in the _sort param values list)
* moved "_sort + _include/_revinclude" check into parseQueryParameters
method where we have all the parameters

2. Enable the new Abstract PL tests.

Removed the commented out classes in testng.xml and added new section
for the
following tests:
* com.ibm.fhir.persistence.jdbc.test.FHIRDBDAOTest
* com.ibm.fhir.persistence.jdbc.test.JDBCNormDeleteTest
* com.ibm.fhir.persistence.jdbc.test.JDBCNormCompartmentTest
* com.ibm.fhir.persistence.jdbc.test.JDBCNormMultiResourceTest
* com.ibm.fhir.persistence.jdbc.test.JDBCNormSortTest
* com.ibm.fhir.persistence.jdbc.test.JDBCNormWholeSystemSearchTest

3. Get everything passing

Signed-off-by: lmsurpre <lmsurpre@us.ibm.com>