issue #92 - FHIRModelTestBase becomes TestUtil
Our test class hierarchy was getting a bit out-of-hand.

Most of the methods in this class were already static, and there was
really nothing gained from the class hiearchy (other than simple
invocation of shared logical functions), so I turned it into a utility.

Hopefully removing this one layer helps a bit.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>