issue #93 - refactor JDBCNormalizedQueryBuilder and rename classes
1. Merged AbstractJDBCQueryBuilder into JDBCNormalizedQueryBuilder (it
used to be a parent class for both the "basic" and "normalized" schema,
but now that we only have the "normalized" one, it doesn't make sense to
have this layer.

2. Removed all "Basic" implementation classes and removed "Norm" or
"Normalized" from the remaining classes. For example:
* ParameterNormalizedDAO merged into ParameterDAO and ParameterBasicDAO
is now deleted.
* ParameterDAONormalizedImpl merged into ParameterDAOImpl and
ParameterDAOBasicImpl is now deleted.
* ResourceNormalizedDAO merged into ResourceDAO and ResourceBasicDAO is
now deleted.
* ResourceDAONormalizedImpl merged into ResourceDAOImpl and
ResourceDAOBasicImpl is now deleted.
* FHIRDbDAONormalizedImpl merged into FHIRDbDAO and FHIRDbDAOBasicImpl
is now deleted.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>