issue #305 - improve datetime processing in JDBCQueryBuilder
1. Make AbstractWholeSystemSearchTest extend AbstractPLSearchTest and
update the tests to verify the expected resource is actually returned
from the search results (previously it only verified that we had some
results).

2. Reduce log level for cases where we have a search parameter and
either A. its not found in the
extension-search-parameter-valuetypes.json config file or B. the
extension-search-parameter-valuetypes.json config file isn't found at
all for the tenant. Previously, I thought these were error conditions,
but now I realize that 'A' is actually the most common path...whenever
we get a search for a built-in search parameter we'll hit this clause.

3. Remove saved resources and reset the tenant after AbstractSortTest.
Previously, this test set the tenant but never restored it afterwards.

4. Fix JDBCQueryBuilder to use the Instant that was created from the
search parameter query value's DateTime (padded if needed). This means
we need to convert it to a java.sql.Timestamp in order to bind it to the
query since JDBC can't auto-convert from java.util.Instant for some
reason.  Additionally, I relaxed the generic type on SqlQueryData so
that we can be more type-safe when adding bindVariables.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>