issue #314 - introduce PL tests for paging and include/revinclude
1. introduce PagingTest (Abstract and JDBC) to verify PL paging behavior
for both search and history interactions

2. re-introduce IncludeRevincludeTest (Abstract and JDBC) to verify
_include/_revinclude behavior. we previously had
AbstractQueryIncludeTest but we had commented it out in our transition
to R4 and then I deleted it as part of pull request #279. this PR
revives this old test and updates it to align with our new testing
patterns.

3. added a single _revinclude test for each search parameter type to
test the combination of search parameters and the _revinclude parameter

4. minor cleanup in FHIRResource to highlight potential issue if we have
more resources than we can fit in Integer.MAX_VALUE (>2 billion)

5. minor doc updates to Conformance.md

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>