remove the profile for java11
it still builds with OpenJDK 11, it just cross-compiles for java 1.8.

i kept the java.version property, so folks can still pass `-Djava.version=11` if they really want to compile for java 11, but since we want to be buildable and runnable on java 8 we won't be using any java 11 source features at this time.

also set fail-fast to false so that we get both java 8 and java 11 results each time

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>