issue #355 - assert baseType is not null before recursing
In the currently implementation, FHIRPathType.isAssignableFrom will
always return true.

For example, if we check is FHIR_DATE_TIME assignable from FHIR_BOOLEAN,
it will recurse until finally we had a baseType of null:
* is FHIR_DATE_TIME assignable from FHIR_ELEMENT
* is FHIR_DATE_TIME assignable from FHIR_ANY
* is FHIR_DATE_TIME assignable from null

And, since everything is assignable from null, it will always be true.

This PR changes that by returning false when we get to baseType == null.

Additionally:
* removal of jdk11 profile for fhir-examples (to match earlier change
with rest of project)
* a logging improvement in
FHIRPersistenceJDBCImpl.extractSearchParameters and
* stricter implementation of `as` wrt arguments that cannot be resolved
to type identifier.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>