issue #355 - introduce FHIRPath FHIR_UNKNOWN_RESOURCE_TYPE
This is a special resource type to use from the resolve() function so
that we will index references to resources for which we cannot determine
the type (e.g. absolute URLs that we either choose not to dereference or
CANNOT be dereferenced).

Previously we used the FHIR_ANY type for this, but issue #355 was
covering an issue related to this.  Specifically the semantics we want
for resolve() are this:
* UNKNOWN as X    X is assignable from UNKNOWN    TRUE
* X as UNKNOWN    UNKNOWN is assignable from X    FALSE

Whereas with ANY, what we implemented was the opposite (ANY is
assignable from X)

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>