issue #374 - introduce FHIRPathSpecTest and fix minor evaluator issues
1. downloaded FHIRPath spec tests from
http://build.fhir.org/ig/HL7/FHIRPath/branches/master/N1/tests.html and
added to fhir-model/src/test/resources

2. tweaked tests-fhir-r4.xml so it could be read successfully (TODO:
open GForge ticket or pull request back to FHIRPath repo with same
changes)

3. introduced FHIRPathSpecTest for reading the XML test file, executing
the expressions, and comparing the actual vs expected output

4. updated =, ~, !=, and !~ operators to return empty if either of the
operands are empty

5. updated `/` operand to catch divide by zero ArithmeticException and
return empty in this case

For now, the test will only fail in case of an unexpected error.
Instead, I'm just counting the number of failures and printing it at the
end.
We're currently passing about 368 of the 686 FHIRPath tests and I
think most of the failures are from unsupported operations
(java.lang.UnsupportedOperationException).

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>