issue #374 - refactored again into TestNG DateProvider factory
The separate FHIRPathSpecTestFactory was working well when invoked from
eclipse, but when executed from the maven build I found that:
1. surefire tries to invoke the test (which should only be invoked by
the factory); and is NOT invoking the factory
2. surefire does NOT invoke the factory class, so just excluding
FHIRPathSpecTest wasn't sufficient

Now we stuff everything into FHIRPathSpecTest and we use the
dataProvider argument so that TestNG will know where to get the test
data from.

Additionally, I added logic to set the status to SKIP for tests which
fail due to unsupported operations or functions (by throwing a
SkipException).  Now we're at:
> Total tests run: 686, Failures: 166, Skips: 150

Finally, I've moved the "catch ArithmeticException" (divide-by-0)
handling outside the arithmetic operators switch statement so we have
consistent behavior across these operators.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>