issue #418 - add order by clause to searchByIds to ensure sorted results
1. Updated the SearchPerformanceTest so that the observations have
components with values...just to easily create many observations which
helps verify the search behavior.  This change could be removed if we
want.

2. Deleted unused variants of searchForIds and searchByIds. I think
these came from the original "basic" schema that was removed for
https://github.com/IBM/FHIR/issues/93 but they were no longer in use and
should be safe to delete.

3. Introduced ORDER BY CASE clause to ResourceDAOImpl.searchByIds.
I found that DERBY is always returning the results in the same order as
the resource ids provided in the "IN" clause, but in general we
shouldn't rely on that.  I tested the CASE statements on derby with up
to 1000 results, but we need to run a similar test on Db2 to ensure it
works there as well.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>