Generate inline primitive types
In the previous version, we generated a definition for each FHIR
primitive type. This correctly captured the fact that these can have an
id and/or extensions (inhereted from Element) but would lead to the
wrong JSON since these elements need to get serialized differently in
JSON.

Technically, we should introduce a `_fieldname` variant for each
instance of a primitive type, but primitive extensions are not common
and it adds significant complexity to the swagger.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>