Add validation for Element.id and Extension.url
These elements are "special" in that, in the XML serialization, they are
attributes and not elements.
To facilitate this "specialness", we keep them as raw java strings in
our model, but according to the spec they should still be valid data
types (String and Id respectively).

Also added javadoc for ValidationSupport.java

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>