more swagger updates
1. I found that the swagger definitions are missing the 'resourceType'
propery; after this change it should be there for all resource types.
  Unfortunately, I think we need to change our whole approach if we want
to indicate the value is 'fixed'. Because there are some places where
any Resource is allowed, I wanted to show that the resourceType field is
an enum and not just String. However, Swagger/OpenAPI--or at least the
Liberty support for it--doesn't like it when a property is defined in
both halves of the *allOf* constraint...even if those two are compatib
le. To address it, I think we should avoid the *allOf* construct and
just copy the duplicated Resource properties into each different
Resource subtype. For now, its just a TODO in the code.

2. Added an extra check to prevent the creation of a DomainResource
definition...only children of DomainResource should get an endpoint, not
DomainResource itself.

3. Introduced generateAllInOne for the FHIROpenApiGenerator. When using
an API Gateway like IBM API Connect, we found it most useful to generate
one swagger definition per resource type. However, I couldn't figure out
how to make Liberty work with multiple static files, and so instead I
created this new method. By default, both methods will be called from
Main and the 'all-in-one' will be written to a file named
"all-openapi.json". I did *not* create a corresponding method for the
Swagger (2.0) variant.

4. Used the new method to regenerate our openapi.json file thats
included in the fhir-openapi web project.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>