issue #460 - refactor R4ExamplesDriver to invoke tests as appropriate
1. Introduced Index.java in the fhir-examples module and moved the
"TestType" concept from fhir-model into this class as a list of
filenames with corresponding Enum values.

2. Renamed ExamplesUtil.reader to "resourceReader" and introduced a
separate helper for getting a Reader from an Index instance

3. Moved all fhir-example index files to the top level of the
fhir-examples module and updated the R4ExampleDriver logic to infer the
resource type from the filename's file extension

4. Moved System Property logic from R4ExamplesDriver to the classes
which call this driver, making it easier to set each one individually

5. Updated pom.xml files and our CI's build.yml so that fhir-model,
fhir-validation, fhir-search, and fhir-persistence-jdbc tests can all be
configured independently (and automatically based on the changes in the
pull request).

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>