throw IllegalStateException instead of Error
I found that the FHIRValidator was throwing an Error instance when a
constraint could not be evaluated.
This seems wrong to me since Error indicates a "serious problems that a
reasonable application should not try to catch".
Futher, this was causing a test failure because the R4ExamplesDriver is
expecting an Exception for invalid examples, not an Error.

Someone should still look into whether a FHIRPath operation (like `&`)
should work over FHIR nodes that have no value (only extensions).  But
at least now all our tests should be passing.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>