issue #403 - Added 4.0.1 definitions and updated CodeGenerator and model
Updated in both fhir-tools and fhir-registry. Next I need to do
fhir-search.

In 4.0.1, they updated the way simple/system types are represented in
the ElementDefinitions. We were already using java.lang.String for
Element.id and Extension.url, but now we use this for Resource.id as
well.  This had a lot of knock-on effects, but overall its a nice
simplification to the model.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>