issue #403 - regenerated fhir-examples for 4.0.1
The biggest change is that Resource.id is no longer populated.
4.0.1 clarified that this element cannot have and id or extensions and
so `_id` is no longer a valid attribute name.

I also needed up update the MinimalDataCreator to add the tag before the
objects are built because now the "all elements must have a value or
children" applies to Resources too and it is checked during object
construction (previously we added these tags only at the very end).

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>