issue #403 - add rootResource external constant
FHIR 4.0.1 introduces the notion of %rootResource and tightens up the
definition of %resource.

Previously we needed to infer what the value of %resource should be
based on the type on which the constraint was found.

Now the $rootResource constant will now be set alongside the $resource
constant for each constraint getting validated from FHIRValidator.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>