issue #403 - Set path for non-value system string nodes in PathAwareVisitor
Added visit(String) method in PathAwareVisitor to make up for the fact
we still want Resource.id to be the path while visiting the Resource.id
java.util.String (e.g. for ResourceFingerprintVisitor). I believe this
is now slightly more correct for the other java.util.String members in
the model as well.

Added hacky XMLGenerator workaround for XMLStreamWriter issue where it
fails to write the end of the last self-closing (aka empty) element tags
(TODO: open issue or get to bottom of that).

As a side-effect of this change, the toString representation of a FHIR
primitive element will now include the type name as well as the value. I
slightly preferred the previous behavior, but this way is more
consistent with the XML toString and I didn't think it worth the effort
to preserve the old JSON toString behavior.

As a result of this side-effect, JDBCIncludeRevincludeTest was failing,
so I made the simple change to get the Id value instead of relying on
the toString behavior there.

Updated fhir-model ToStringTest to verify toString works as expected.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>