Added debug capabilities to IsDistinctFunction
Now its possible to see which nodes are not distinct...in my case there
were 5 repeats in a huge spec file and it was impossible to find without
this.

Also added trace line to ValidationProcessor to make it possible to see
the full detail of the Issues in case of validation failure.

Also updated our version of testng, because why not.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>