issue #317 - update SearchUtil and JDBC PL to handle composite search
Also updated BasicReference.json so that Reference-id has an identifier
and no reference value. Previously I thought that a reference value
could be just a resource id with the type prefix, but it turns out this
"infer the type" behavior is only for the query and not the resource
value.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>