issue #317 - updated add_any_resource stored proc
it turns out we don't pass the parameters into the stored proc any more,
so the only update here was to delete the parameters from the composites
when we get a new version of the resource (which covers both `delete`
and `update` cases).

also, i deleted the stale `addParameterArrayTypes.sql` stored procedure
template, and the corresponding methods that were defining custom array
types that were only used from this one procedure.

finally, i added a dependency from the resource tables to the "global"
search parameter values tables in fhir-persistence-schema...this to
avoid a deadlock I was hitting very frequently while attempting to
deploy the schema to db2 in a local container

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>