Update Db2Adapter.createFhirSchemas to succeed when schemas exist
Without this, the step always fails if you've already created a schema
on this db because FHIR_ADMIN already exists. Now it will succeed but
print a WARNING to indicate it already existed.

1. introduced DuplicateNameException and a corresponding helper to
identify when schema creation fails due to the schema already existing

2. renamed DuplicateValueException to UniquenessViolationException to
avoid confusion with DuplicateNameException

3. updated Db2Adapter to catch the DuplicateNameException and WARN
instead of failing

Also added log statement at the end of tenant creation so that you don't
need to scroll to the very top of the output to find the generated key

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>