issue #519 - QueryParameterValue.toString should use query string format
I didn't realize that we were using QueryParameterValue.toString to
write the self URI for the response messages, so I added a generic
toString capability for composite values.
Now, I've updated this to return the data in a format that is consistent
with the query string value we parsed to get it.

Note: there is no known syntax for a single component within a composite
value to itself have multiple values; we'll throw an
IllegalStateException if we ever see it.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>