issue #564 - package the swagger-generator in the distribution
Now the executable `fhir-swagger-generator` jar will be in the
`INSTALL_DIR/wlp/fhir` directory.

I also modified the packaging for the fhir-persistence-schema executable
to be in the same place.
Previously, we set a finalName that excluded the version in the
filename. Now the shaded jar will include the version and get a custom
"cli" classifier and this is the version that will be packaged alongside
the fhir-swagger-generator.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>