issue #608 - introduce the ALL index and use it for fhir-model changes
Also refactored the Index class to take a list of paths so that we don't
need to maintain the same rows in 3 different files.
Maintaining backwards compat was a little ugly, but the index files
shouldn't be too big and so combining them into a single string
shouldn't be too bad.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>