Remove hacky logic for retrieving the element name from a path
It turns out that our FHIRPathNode object already exposes the element
name from a selected node.
This works for all cases other than inserting an element at position 0
of an empty list, and its not clear from the spec whether that should be
supported or not. Now we throw UnsupportedOperationException in that
case.

I also reduced the visibility of all the patch visitors, added null
checks, and improved documentation of both the visitors and a few select
methods in the FHIRPath model.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>