issue #679 - fixed major issue caused by use of parent identity
The previous approach failed in certain cases when the Visitable tree
contains the exact same object in multiple different places (e.g. adding
to a name when that same name object is repeated twice in teh same
list).

With this change, the CopyingVisitor now keeps its own pathStack and we
use the normalized path of FHIRPath nodes and/or their parent to ensure
we're only modifying the proper paths in the tree.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>