issue #706 - add low and high values for all numeric search params
1. Introduce NUMBER_VALUE_LOW and NUMBER_VALUE_HIGH columns to all
X_NUMBER_VALUES tables

2. Update JDBCParameterBuildingVisitor and ParameterVisitorBatchDAO to
insert to these for all extracted numbers. For integers, we insert the
exact value whereas for decimals we compute the implicit range.
Similarly, we will now populate the QUANTITY_VALUE_LOW and
QUANTITY_VALUE_HIGH columns for Quantities as well as Ranges.

Gap:  a quantity value has an exclusive upper bound whereas a Range has
an inclusive upper bound. Presently, the only way to distinguish these
cases is that the Quantity will have a value in the QUANTITY_VALUE
column as well (whereas Ranges will not).

3. Move all numeric behavior from QuantityParmBehaviorUtil to
NumberParmBehaviorUtil and update it to use ranges...now we'll have a
unified approach across integers, decimals, quantities, and Range.

4. Update all tests

With these changes, we will now use the implicit range value of all
numeric values, even when comparator prefixes are specified on the
search.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>