issue #17 - move exception handling and logging into FHIRResource
Previously, we logged all errors in the `doX` methods which are now in
FHIRRestHelper. However, this can result in an issue if there is an
exception that occurs after the `doX` method but before we return the
response.

Similarly, before the prior commit, we were writing audit log messages
from the `doX` methods as well.

With this change
1. all exception handling is moved to FHIRResource
2. exception logging is consolidated under
`FHIRResource.exceptionResponse`
3. audit logging is added to a finally block for each JAX-RS operation

I also
* tweaked FHIRResourceHelpers.doBundle to take a bundle and moved that
validation up to FHIRResource
* updated SearchExceptionUtil to include OperationOutcome.Issue so that
we don't need to special-case search exceptions
Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>