issue #700 - intelligently set reference types based on constraints
1. Introduced `com.ibm.fhir.model.annotation.ReferenceTarget` for
capturing the allowed target types for each Reference type in the model

2. Updated CodeGenerator to include these new annotations and also add
the allowed reference types to the javadoc of the Builder setters

3. Added support for introspecting this data in ModelSupport...similar
to what we did for allowed Choice types in the past

4. Update ExamplesGenerator to set the `type` of each Reference based on
the allowed type and prefix the reference value itself with the same
type

For number 4, I re-used the `choiceIndicator` to pick from among the
allowed reference types.

Finally, I updated the ExamplesGenerator to use ModelSupport for
choiceElement information and removed the prior approach of using a
pre-computed properties file (choiceElements.properties).

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>
Signed-off-by: ccorley <ccorley@us.ibm.com>