issue #781 - skip constraint checking for composite foreign keys
We found that the query plan was doing a full index scan of the
Observation_COMPOSITES table on each delete to each of the
Observation_X_VALUES tables. This is totally unneccessary because we
always delete the composite rows first before we delete the
corresponding. Telling Db2 not to enforce this particular constraint
should speed things up nicely.

Also fixed javadoc in FhirResourceTableGroup by:
1. fixing copy-paste issue in addComposites javadoc
2. remove "ON DELETE CASCADE" from everywhere --its simply not used
3. added "NOT ENFORCED" for the addComposites sample FKs

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>