issue #833 - add migration logic for #706 and more
Previously, I forgot we made this schema change early in the 4.1.0 cycle
and so the migration test was starting with a schema that didn't exactly
match 4.0.1.

To make migration match the schema exactly (same columns in the same
order), I had to move these two "new" number columns from the middle of
the table definition to the end.

Additionally, I found that FhirSchemaGenerator was still including the
retired `DATE_VALUES` column in the global/resource-level DATES table.
While that table isn't used, I updated it to match the resource-specific
ones.

Also added an `@implNote` to PopulateResourceTypes per review comment.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>
Signed-off-by: ccorley <ccorley@us.ibm.com>