Generate examples for Element.extension and Identifier.assigner
We found that certain Swagger/OpenAPI tooling was attempting to generate
examples if none were provided in the Swagger/OpenAPI definitions.

A more complete fix would be to generate our own examples for all FHIR
data types, but this should be sufficient for avoiding the
MAX_DEPTH_ERROR we've been seeing in API Connect.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>