issue #833 - add db2 migration test and fix issues
Introduced variant of e2e-db2 test for testing migration which:
1. creates a db2 instance and deploys the previous version of the schema
2. runs migration
3. executes the e2e tests

This uncovered a couple different issues with our existing migrations:
1. `DB2 SQL Error: SQLCODE=-2310, SQLSTATE=     , SQLERRMC=-668`
    while executing runstats after dropping a column
2. `DB2 SQL Error: SQLCODE=-108, SQLSTATE=42601,
SQLERRMC=FK_OBSERVATION_COMPOSITES_DATE`
    while executing DropForeignKeyConstraint
3. `DB2 SQL Error: SQLCODE=-204, SQLSTATE=42704,
SQLERRMC=FK_OBSERVATION_COMPOSITES_DATE`
    while executing DropForeignKeyConstraint

For number 1, the fix was to avoid executing RUNSTATS until after the
table with the dropped column has been REORGed.

For number 2, the fix was to stop qualifying the constraintName with the
schema name.

For number 3, the fix was to avoid hard-coding Observation in the FK
names being dropped.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>