issue #910 - introduce fhir-context and move the context classes there
* FHIRRequestContext
* FHIRPagingContext
* FHIRHistoryContext

For now, I kept FHIRSearchContext in fhir-search and kept that mostly
separate from this new module.

As a result of the change, I had to remove the tenantId lookup from
FHIRConfigurationHelper since it no longer has access to the context
classes.  Instead, I just added a string "tenantId" parameter to all the
get methods and do the tenantId lookup before calling the methods.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>