Move Liberty OAuth tables to the FHIR_OAUTH schema and gen by default
Previously, a user needed to create these tables themselves if they
wanted to use liberty as an OAuth 2.0 provider with a databaseStore
(e.g. to dynamically manage OAuth 2.0 clients).

Now these tables will be created/updated as part of the
fhir-persistence-schema create-schema/update-schema actions, or as part
of derby bootstrapping.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>