Add codesystem, valuesets, and structuredefinitions referenced from the referenced extensions
Previously, I included only the extensions themselves. Now, I include
all conformance resources needed to validate those extensions (as well
as the valuesets and codesystems that are provided with binding strength
"example"). The new StructureDefinitions might reference yet more
resources, but this set is sufficient for the purposes of validating
DaVinci PlanNet resources which reference these extensions.

Note:  The `StructureDefinition-vhdir-organization.json` in VHDir Ballot
2 was not valid, so I replaced just that one with the 'latest' from
master, which is still labeled as version `0.2.0`.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>