When searchparameter expression resolves to extension element, use value
While testing with the DaVinci PDEX Plan-Net we noticed that some of the
search parameters weren't getting indexed.
Upon further inspection, it was because the expression led to an
Extension element and not the value of that extension.
I'm still not sure what is "best practice" but I found that the
`mothersMaidenName` search parameter from the base spec has the same
issue and so I think its good for us to handle this.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>