issue #1237 - stop adding duplicate SearchParameters with base Resource
I found that we were adding the Resource-level SearchParameters in two
different places and this was causing some inconsistent behavior.
In this option, I stop adding the Resource search-parameters at the end
of getFilteredBuiltinSearchParameters.

This effectively reverts us to the searchfilter behavior we had in our
DSTU2 implementation; if a filter is present for a given type, then the
common parameters must be explicitly listed as allowed search parameters
as well.

Also: removes redundant getSearchParameterByCodeIfPresent in favor of
`getApplicableSearchParametersMap(resourceType).get(code)`

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>