issue #1237 - stop double-adding inhereted searchparameters
I found that we were adding the Resource-level SearchParameters in two
different places and this was causing some inconsistent behavior.
In this option, I stop adding the Resource search parameters in
ParametersUtil and instead update the methods in SearchUtil to ensure
that these parameters are handled properly.

Also: removes redundant `getApplicableSearchParametersMap(resourceType)`
in favor of `getSearchParameter()` calls which eventually reach
`getSearchParameterByCodeIfPresent()`.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>