issue #1257 - handle codes which conflict across resource types
In https://github.com/IBM/FHIR/pull/1238 I introduced a conflict check
for the ParametersMap, but this is only constructed on a
per-resource-type basis.
With this change, we will now detect conflicts where the same code is
defined for both a specific resource type and the more generic
"Resource" type as well.


Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>