issue #1340 - support bulk export to parquet format
1. Updates to fhir-operation-bulkdata and fhir-bulkimportexport-webapp
to accept and pass alternative export formats (application/fhir+parquet)

2. Introduction of SparkParquetWriter for using Apache Spark to
introspect a list of JSON strings (e.g. FHIR resources), infer a schema,
and write the schema and data into a parquet file. This approach writes
specific files under a single logical file which is actually a
directory, allowing the schema to change over time.

3. Updates to the System ChunkReader and ChunkWriter for parquet;
instead of writing to a buffer in the reader, we simply add to the
context info and return the Resource objects themselves. Then, in the
writer, we always write a file with whatever resources were passed.
This works because Spark is managing the writes to COS for us and
Parquet has this notion of a multi-part logical file, so we don't need
the tight control over multi-part uploads like in the case of NDJSON.

Also includes miscellaneous cleanup, debug, and style changes.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>