Adjust default cosFileMaxResources and skip the flatten step
1. Reduced default cosFileMaxResources from 500,000 to 200,000.
The Parquet implementation will hold all of these objects in memory (as
both objects and as strings) and so a smaller size seems more
appropriate.
Furthermore, even for an NDJSON export with an average size of just
1KB/resource, the 500000 number would lead to 500 MB files whereas the
default cosFileMaxSize is 200 MiB. Therefor, adjusting
cosFileMaxResources to 200,000 is more in line with the other default.

Note: If this still proves too memory-intensive, an alternative approach
is to append each page of resources to a temporary ndjson file (from
within the Reader) and then load that file via
`spark.read().json(tmpFile)` before writing it to COS. Another potential
alternative is to use a MappedByteBuffer (in place by the
TransientUserData bytebuffer) and use that for both NDJSON and Parquet.

2. Remove the "flattenAndClean" step.
Spark 3.0 has made improvements in terms of running analytics on nested
data and so, after discussing with Gidon, we decided that the flatten
step is unnecessary.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>